pyproject.toml
[project]
name = "conduit-coding-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
]
